+++ source/common/router/scoped_rds.cc
@@ -159,7 +159,9 @@ ScopedRdsConfigProvider::ScopedRdsConfigProvider(
           MutableConfigProviderCommonBase::subscription_.get())),
       rds_config_source_(std::move(rds_config_source)) {
   initialize([scope_key_builder](Event::Dispatcher&) -> ThreadLocal::ThreadLocalObjectSharedPtr {
-    return std::make_shared<ThreadLocalScopedConfigImpl>(scope_key_builder);
+    return std::make_shared<ThreadLocalScopedConfigImpl>(
+        envoy::config::filter::network::http_connection_manager::v2::ScopedRoutes::ScopeKeyBuilder(
+            scope_key_builder));
   });
 }