+++ test/server/server_test.cc
@@ -303,6 +303,7 @@ TEST_P(ServerInstanceImplTest, StatsFlushWhenServerIsStillInitializing) {
 
   // Wait till stats are flushed to custom sink and validate that the actual flush happens.
   TestUtility::waitForCounterEq(stats_store_, "stats.flushed", 1, test_time_.timeSystem());
+  EXPECT_EQ(3L, TestUtility::findGauge(stats_store_, "server.state")->value());
   EXPECT_EQ(Init::Manager::State::Initializing, server_->initManager().state());
 
   server_->dispatcher().post([&] { server_->shutdown(); });